adding new classification model
