+++ model_rnn.py
@@ -1,22 +1,23 @@
 import model_data as data_processing
-import model_structure as model_structure
 import model_prediction as model_prediction
 import model_classification as model_classification
 
 # Program header
-print("---  Simple RNN Neural Network                ---")
 print("---  Starting Neural Network                  ---")
 print("-------------------------------------------------")
 
 # Get Data
-trainingData, validateData, testData = data_processing.readData()
-
+# trainingData, validateData, testData = data_processing.readData("kddcup99_csv.csv")
 # # Create Model Prediction
-# model = model_structure.defineModel()
+# model = model_prediction.defineModel()
 # # Train and Validate Model
 # model = model_prediction.trainModel(3, model, trainingData, validateData)
 
+
+# Get Data
+trainingData, validateData, testData = data_processing.readData("kddcup99_csv_balance.csv")
+
 # Create Model Classification
-model = model_structure.defineModelClassification()
+model = model_classification.defineModel()
 # Train and Validate Model
 model = model_classification.trainModel(3, model, trainingData, validateData)
