+++ references.txt
@@ -23,3 +23,4 @@ https://www.mdpi.com/2078-2489/11/5/243/htm
     # PCA 
     # https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html
 
+https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.to_numpy.html
